Move promptPassword to Helper class
