+++ app/src/main/java/com/m2049r/xmrwallet/LoginActivity.java
@@ -29,20 +29,14 @@ import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Handler;
 import android.support.annotation.NonNull;
-import android.support.design.widget.TextInputLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
-import android.support.v4.hardware.fingerprint.FingerprintManagerCompat;
-import android.support.v4.os.CancellationSignal;
-import android.text.Editable;
-import android.text.TextWatcher;
 import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.inputmethod.EditorInfo;
-import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -56,7 +50,6 @@ import com.m2049r.xmrwallet.model.NetworkType;
 import com.m2049r.xmrwallet.model.Wallet;
 import com.m2049r.xmrwallet.model.WalletManager;
 import com.m2049r.xmrwallet.service.WalletService;
-import com.m2049r.xmrwallet.util.CrazyPassEncoder;
 import com.m2049r.xmrwallet.util.FingerprintHelper;
 import com.m2049r.xmrwallet.util.Helper;
 import com.m2049r.xmrwallet.util.KeyStoreHelper;
@@ -183,7 +176,7 @@ public class LoginActivity extends SecureActivity
                     case DialogInterface.BUTTON_POSITIVE:
                         final File walletFile = Helper.getWalletFile(LoginActivity.this, walletName);
                         if (WalletManager.getInstance().walletExists(walletFile)) {
-                            promptPassword(walletName, new PasswordAction() {
+                            Helper.promptPassword(LoginActivity.this, walletName, new Helper.PasswordAction() {
                                 @Override
                                 public void action(String walletName, String password) {
                                     startDetails(walletFile, password, GenerateReviewFragment.VIEW_TYPE_DETAILS);
@@ -215,7 +208,7 @@ public class LoginActivity extends SecureActivity
         if (checkServiceRunning()) return;
         final File walletFile = Helper.getWalletFile(this, walletName);
         if (WalletManager.getInstance().walletExists(walletFile)) {
-            promptPassword(walletName, new PasswordAction() {
+            Helper.promptPassword(LoginActivity.this, walletName, new Helper.PasswordAction() {
                 @Override
                 public void action(String walletName, String password) {
                     startReceive(walletFile, password);
@@ -473,153 +466,6 @@ public class LoginActivity extends SecureActivity
     public void onAddWallet(String type) {
         if (checkServiceRunning()) return;
         startGenerateFragment(type);
-    }
-
-    AlertDialog openDialog = null; // for preventing opening of multiple dialogs
-
-    void promptPassword(final String wallet, final PasswordAction action) {
-        if (openDialog != null) return; // we are already asking for password
-        Context context = LoginActivity.this;
-        LayoutInflater li = LayoutInflater.from(context);
-        final View promptsView = li.inflate(R.layout.prompt_password, null);
-
-        AlertDialog.Builder alertDialogBuilder = new AlertDialog.Builder(context);
-        alertDialogBuilder.setView(promptsView);
-
-        final TextInputLayout etPassword = (TextInputLayout) promptsView.findViewById(R.id.etPassword);
-        etPassword.setHint(LoginActivity.this.getString(R.string.prompt_password, wallet));
-
-        boolean fingerprintAuthCheck;
-        try {
-            fingerprintAuthCheck = FingerprintHelper.isFingerprintAuthAllowed(wallet);
-        } catch (KeyStoreException ex) {
-            fingerprintAuthCheck = false;
-        }
-
-        final boolean fingerprintAuthAllowed = fingerprintAuthCheck;
-        final CancellationSignal cancelSignal = new CancellationSignal();
-
-        if (fingerprintAuthAllowed) {
-            promptsView.findViewById(R.id.txtFingerprintAuth).setVisibility(View.VISIBLE);
-        }
-
-        etPassword.getEditText().addTextChangedListener(new TextWatcher() {
-
-            @Override
-            public void afterTextChanged(Editable s) {
-                if (etPassword.getError() != null) {
-                    etPassword.setError(null);
-                }
-            }
-
-            @Override
-            public void beforeTextChanged(CharSequence s, int start,
-                                          int count, int after) {
-            }
-
-            @Override
-            public void onTextChanged(CharSequence s, int start,
-                                      int before, int count) {
-            }
-        });
-
-        // set dialog message
-        alertDialogBuilder
-                .setCancelable(false)
-                .setPositiveButton(getString(R.string.label_ok), null)
-                .setNegativeButton(getString(R.string.label_cancel),
-                        new DialogInterface.OnClickListener() {
-                            public void onClick(DialogInterface dialog, int id) {
-                                Helper.hideKeyboardAlways(LoginActivity.this);
-                                cancelSignal.cancel();
-                                dialog.cancel();
-                                openDialog = null;
-                            }
-                        });
-        openDialog = alertDialogBuilder.create();
-
-        final FingerprintManagerCompat.AuthenticationCallback fingerprintAuthCallback = new FingerprintManagerCompat.AuthenticationCallback() {
-            @Override
-            public void onAuthenticationError(int errMsgId, CharSequence errString) {
-                ((TextView) promptsView.findViewById(R.id.txtFingerprintAuth)).setText(errString);
-            }
-
-            @Override
-            public void onAuthenticationSucceeded(FingerprintManagerCompat.AuthenticationResult result) {
-                String userPass = KeyStoreHelper.loadWalletUserPass(LoginActivity.this, wallet);
-                if (processPasswordEntry(wallet, userPass, action)) {
-                    Helper.hideKeyboardAlways(LoginActivity.this);
-                    openDialog.dismiss();
-                    openDialog = null;
-                } else {
-                    etPassword.setError(getString(R.string.bad_password));
-                }
-            }
-
-            @Override
-            public void onAuthenticationFailed() {
-                ((TextView) promptsView.findViewById(R.id.txtFingerprintAuth))
-                        .setText(getString(R.string.bad_fingerprint));
-            }
-        };
-
-        openDialog.setOnShowListener(new DialogInterface.OnShowListener() {
-            @Override
-            public void onShow(DialogInterface dialog) {
-                if (fingerprintAuthAllowed) {
-                    FingerprintHelper.authenticate(LoginActivity.this, cancelSignal, fingerprintAuthCallback);
-                }
-                Button button = ((AlertDialog) dialog).getButton(AlertDialog.BUTTON_POSITIVE);
-                button.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        String pass = etPassword.getEditText().getText().toString();
-                        if (processPasswordEntry(wallet, pass, action)) {
-                            Helper.hideKeyboardAlways(LoginActivity.this);
-                            openDialog.dismiss();
-                            openDialog = null;
-                        } else {
-                            etPassword.setError(getString(R.string.bad_password));
-                        }
-                    }
-                });
-            }
-        });
-
-        // accept keyboard "ok"
-        etPassword.getEditText().setOnEditorActionListener(new TextView.OnEditorActionListener() {
-            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                if ((event != null && (event.getKeyCode() == KeyEvent.KEYCODE_ENTER)) || (actionId == EditorInfo.IME_ACTION_DONE)) {
-                    String pass = etPassword.getEditText().getText().toString();
-                    if (processPasswordEntry(wallet, pass, action)) {
-                        Helper.hideKeyboardAlways(LoginActivity.this);
-                        openDialog.dismiss();
-                        openDialog = null;
-                    } else {
-                        etPassword.setError(getString(R.string.bad_password));
-                    }
-                    return true;
-                }
-                return false;
-            }
-        });
-
-        Helper.showKeyboard(openDialog);
-        openDialog.show();
-    }
-
-    interface PasswordAction {
-        void action(String walletName, String password);
-    }
-
-    private boolean processPasswordEntry(String walletName, String pass, PasswordAction action) {
-        String walletPassword = Helper.getWalletPassword(getApplicationContext(), walletName, pass);
-        if (walletPassword != null) {
-            action.action(walletName, walletPassword);
-            return true;
-        } else {
-            return false;
-        }
     }
 
     ////////////////////////////////////////
@@ -1251,7 +1097,7 @@ public class LoginActivity extends SecureActivity
         File walletFile = Helper.getWalletFile(this, walletNode.getName());
         if (WalletManager.getInstance().walletExists(walletFile)) {
             WalletManager.getInstance().setDaemon(walletNode);
-            promptPassword(walletNode.getName(), new PasswordAction() {
+            Helper.promptPassword(LoginActivity.this, walletNode.getName(), new Helper.PasswordAction() {
                 @Override
                 public void action(String walletName, String password) {
                     startWallet(walletName, password);
